2.0.1
Woohoo. Why already 2.0.1? Because this update is fucking massive. Adds settings where you can turn on the languages, the filled in text is not green so you can clearly see what you filled in and there's no need for the "Selected cards" list. AND LIKE 20* BIGGER THAN ALL OF THIS: The background code is finally not a fucking peace of shit. Say goodbye to having all 32k cards loaded in the background even tho you're playing with a pack of 600. And yeah, that's probably it. Oh yeah, the code is still hot garbage. No problem.

One issue is when you start the game, you can for some time see the loading screen and it kinda annoys me but whatever.