2.1
Done:
1. Added Catalan language
2. Added an option to get x random packs chosen for you. You can pick precisely how many packs get picked and you can press a reload button to regenerate the packs. You have no way of knowing which ones are selected so far, so you have to do it just based on the card amount.
3. The app is now again showing you how many cards are in selected packs which was removed because of a change in how the screen model worked - previous update.

TODO:

If i want to add more content to the app, it will probably be a multiplayer - probably still on one device, because I'm really trying to avoid all that networking bullshit. It would work as the original game - the players take turns being the "Tzar" each round. The other players complete the sentence from their cards and then the Tzar awards a point to the best completion.

I tried to find a way to add some on device AI which would judge how good the completion is, but even the biggest models struggle with humor, so the models who can run on a phone... probably wouldn't really work well.